st_app/requirements.txt
Pillow==11.2.1
Requests==2.32.4
streamlit==1.45.1
boto3==1.40.0